Tests unitaires de persistence

+++ src/main/java/fr/sopra/pox3/ejb/MaisonDeDisqueDAO.java
@@ -1,5 +1,6 @@
 package fr.sopra.pox3.ejb;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.ejb.Stateless;
@@ -7,6 +8,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 
+import fr.sopra.pox3.entities.Auteur;
 import fr.sopra.pox3.entities.MaisonDeDisque;
 
 @Stateless
@@ -33,15 +35,25 @@ public class MaisonDeDisqueDAO {
 		em.persist(maison);
 	}
 	
-	public void deleteMaisonDeDisque (MaisonDeDisque maison) {
-		em.remove(em.contains(maison) ? maison : em.merge(maison));
+	public void deleteMaisonDeDisque (MaisonDeDisque maison) throws Exception {
+		em.merge(maison);
+		TypedQuery<Auteur> q = em.createQuery("from auteurs ", Auteur.class);
+		List<Auteur> auteurs = new ArrayList<>();
+		for (Auteur auteur : q.getResultList()) {
+			if (auteur.getMaison().getId() == maison.getId()) {
+				auteurs.add(auteur);
+			}
+		}
+		
+		if (auteurs.isEmpty()) {
+			em.remove(em.contains(maison) ? maison : em.merge(maison));
+		} /*else {
+			throw  new Exception("Impossible d'effacer une maison contenant des auteurs");
+		}*/
 	}
 	
 	public void updateMaisonDeDisque (MaisonDeDisque maison) {
 		em.merge(maison);
-		/*MaisonDeDisque maisonAUpdate = this.findById(maison.getId());
-		MaisonDeDisque maisonUpdated = em.merge(maisonAUpdate);
-		maisonUpdated.setNom(maison.getNom());*/
 	}
 	
 }
